Fix to switching current extraction and voltage offset correction
Since the DMM has an offset, the superconducting region of a VI curve isn't exactly at zero. When extracting the switching current this makes it hard to get the switching current to read zero at the nodes of a Fraunhofer for example (if the nodes do reach zero) unless you had the threshold set exactly right. This can mess with the current distribution extraction and make it unstable for slight changes in thresholds especially since the threshold isn't the same for all measurements . To fix this, the voltage offset is corrected(which also needed a quick fix) first before extracting the switching current. The threshold should hover now around zero and the threshold should be set to a small value (basically a voltage range for which a point in the superconducting region can vary around zero before not being considered superconducting anymore) which could be set 3e-5 which seems to work for all measurements. Take note if this values doesn't seem to be accurate for specific measurements.