Fixup logging
Logging configuration should be done on the package logger, not the root
logger.

Get the actual log message; record.msg might not be a string (e.g.
matplotlib sometimes passes a LayoutBox) or might be a pre-populated
format string.

Allow configure_logging to be imported as
```python
from shabanipy.logging import configure_logging
```