Support falling back to instrument config
This is useful if you want to match/classify based on e.g. gate voltage,
but the gate voltage wasn't explicitly added to the step list.

If a StepPattern fails to match any explicit steps in the step list and
a fallback InstrumentConfigPattern was specified, the InstrumentConfig
will be matched against the fallback pattern.