Bug fix to data classifying code
Sometimes labber doesn't store the actual data for step fields that aren't iterated on so it doesn't store in the channel names. But the field name is still there in the Step config. When the data_classifying code tries to find the data for each step config name it crashes cause that data wasn't stored. The fix is just an if statement that only tries to find the data if the step config field is also in the channel list.