Revert "Bug fix to data classifying code"
This reverts commit 2e0ce2ed1b169518cc3c90c2a75ef2874b679eef.

This bugfix doesn't work if we're matching on a pattern that's not used
in classification.  In that case, `name` will be reassigned to the
pattern.name (not the StepConfig name), and the if-statement fails even
though there is data to pull.

It's also a bad idea to give the data classifying code a direct
reference to the Labber hdf5 file.  All those interactions should be
encapsulated in LabberData to keep the code modular and not strongly
coupled.