Fit SQUID oscillations to two-JJ CPR model
This model neglects any modulation by a fraunhofer envelope.
Only the positive branch is fitted for now.
There is currently an ambiguity in the handedness of the system when
comparing a single-JJ CPR model (where tilt is in one direction) vs.
two-JJ CPR model (where tilt direction is determined by which Ic
dominates).