Get analyse_vi_curve() tests running (but not passing)
The function signature changed at c0264067 where it seems the intention
was to automatically correct for the voltage offset.

The tests were subsequently added at aebc5dd8 based on the old function
signature.

The implementation has since changed in a series of commits and is
causing the critical current to degenerate to 0 due to 0 peak size.