Exclude relational channels from ivar channels
These are channels that are stepped according to an equation that
relates them to other step channels (via ``Enable channel relations'' in
the step configuration panel).  They are not independently varied but
rather are set depending on the values of other step channels.  More
importantly, they do not contribute an extra dimension to the data,
therefore they are excluded.