Bugfix: Get step channel axis from Channel and squeeze
These bugs affect filtering n-dimensional data.

Step channel axes are shifted by 1 when a trace channel is present.  The
axis in the step config does not account for this; it only considers
channels in the step sequence.

When a filter reduces a dimension to length 1, (e.g. when filtering for
equality), that dimension should be squeezed out.