Pad interrupted scans with NaN's
This will fix most data reshaping errors which continue to occur due to
the inconsistent ways in which Labber handles interrupted steps and
traces.  The tradeoff is memory efficiency.

Reshaping errors may still occur e.g. when filtering interrupted scans.
These can be fixed by generating the would-be step data from the Step
Items in the Step Config.  This fix is deferred until it proves useful
enough to implement.  For now we prefer to only deal with data that was
actually recorded to avoid giving the impression that the generated
would-be data is actually real.