Revert "Invert current to conform to diode sign convention"
This reverts commit 28f63b83d58cfc60b51eb995ab9bf1ef5a6e357d.

This is slightly more complicated for d.c. data as the "voltage probes"
should also be swapped, but the d.c. voltage offset of the amplifier
makes inverting the "raw" voltage nontrivial.

Inverting the critical currents to conform to diode sign conventions
will be done instead in a post-processing step.